fix(api): WIP

+++ api/src/gym/gym.controller.ts
@@ -1,8 +1,20 @@
-import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  UnauthorizedException,
+  UseGuards,
+} from '@nestjs/common';
 import { GymService } from './gym.service';
 import { CreateGymDto } from './dto/create-gym.dto';
 import { RouteSetterService } from '../route-setter/route-setter.service';
 import { RouteSetterDto } from '../route-setter/route-setter.dto';
+import { JwtAuthGuard } from '../auth/auth.guard';
+import { Credentials } from '../auth/credentials.decorator';
+import { Person } from '../person/person.entity';
 
 @Controller('gym')
 export class GymController {
@@ -31,13 +43,27 @@ export class GymController {
     return this.gymService.remove(+id);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Post(':gymId/assign/:personId')
-  assign(@Param('gymId') gymId: string, @Param('personId') personId: string) {
+  assign(
+    @Param('gymId') gymId: string,
+    @Param('personId') personId: string,
+    @Credentials() credentials: Person,
+  ) {
+    if (!credentials?.isAdmin)
+      throw new UnauthorizedException('Not allowed to access this route');
     return this.routeSetterService.create(gymId, personId);
   }
 
   @Delete(':gymId/assign/:personId')
-  unAssign(@Param('gymId') gymId: string, @Param('personId') personId: string) {
+  unAssign(
+    @Param('gymId') gymId: string,
+    @Param('personId') personId: string,
+    @Credentials() credentials: Person,
+  ) {
+    if (!credentials?.isAdmin)
+      throw new UnauthorizedException('Not allowed to access this route');
+
     return this.routeSetterService.remove(gymId, personId);
   }
